[ADD] Item, Skill, Monster, Player 데이터파일 생성 및 GameData 완성 (-성림)
[ADD]
WeGoMars/WeGoMars/
Character.cs 에 Skill 구조체 추가
Item.cs 에 enum ITemType 추가
ItemData, SkillData, MonsterData, PlayerData 의 Json 파일 생성
GameData.cs 초기화 시 위의 Json 파일들에서 데이터를 긁어온다.
(-성림)

+++ WeGoMars/WeGoMars/Monster.cs
@@ -3,7 +3,7 @@ namespace WeGoMars
 {
     public class Monster : Character
     {
-        public Monster(string name, string job, int level, int atk, int def, int maxHp, int maxMp, int hp, int mp, int gold, int exp, List<Skill> skillList, List<Item> inventory)
+        public Monster(string name, string job, int level, float atk, int def, int maxHp, int maxMp, int hp, int mp, int gold, int exp, List<Skill> skillList, List<Item> inventory)
         {
             Name = name;
             Job = job;
